Fix comment end patterns
Ending a comment at ^ rather than $ caused issues with other rules (such as meta.definition.function) that were looking for $ as an end pattern and would only be able to find it on a subsequent line if the comment pattern ate the newline.

I believe this will fix https://github.com/microsoft/vscode/issues/134616.