Remove some conditionals used for regex literals (#17)
Relates to https://github.com/jtbandes/swift-tmlanguage/issues/16 / cc
@slevithan

The conditional negative lookaheads were conditional based on absence of
`#` in the opening delimiter. This can be simplified by including `/` in
the negative lookaheads and moving them before the opening delimiter.

(The comment rule seems like it is not actually necessary since
`#comment` is given a higher priority, but I kept it for good measure.)